+++ packages/workflow/src/nuxt.ts
@@ -1,0 +1,4 @@
+import workflowNuxt from '@workflow/nuxt';
+
+export * from '@workflow/nuxt';
+export default workflowNuxt;