+++ packages/nitro/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @vercel/workflow-nitro
 
+## 4.0.1-beta.12
+
+### Patch Changes
+
+- 98c36f1: Add HMR to nitro integration
+- Updated dependencies [7013f29]
+- Updated dependencies [a28bc37]
+- Updated dependencies [e0c6618]
+- Updated dependencies [809e0fe]
+- Updated dependencies [adf0cfe]
+- Updated dependencies [5c0268b]
+- Updated dependencies [0b3e89e]
+- Updated dependencies [7a47eb8]
+  - @workflow/core@4.0.1-beta.10
+  - @workflow/swc-plugin@4.0.1-beta.4
+  - @workflow/builders@4.0.1-beta.8
+
 ## 4.0.1-beta.11
 
 ### Patch Changes